+++ lib/presentation/pages/auth/auth.dart
@@ -1,5 +1,5 @@
 export 'change_password/change_password_page.dart';
-export 'forgot_password_page.dart';
+export 'forgot_password/forgot_password_page.dart';
 export 'login/login_page.dart';
 export 'otp/otp_page.dart';
 export 'otp/widget/resend_otp_timer.dart';
